GUI improvements.
Added detailed cluster status info in form of modal.

Fixed integrity check & synchronization to not react on "Size" parameter - integrity check will still have text style as changed, but it will not cause synchronization notification & false integrity check to display.

Fixed an issue where two warnings about index displayed when it was missing from the engine side. Now it can be possible, but only when wrong index is registered and does not exists on engine side.